moved functions at the top of index.js to utils/handlers.js
